EvisionWeb/requirements.txt
flask==2.3.2
flask-socketio==5.3.4
opencv-python==4.8.0.76
numpy==1.26.0
google-generativeai==0.3.2
ultralytics
Werkzeug==3.1.3